8.1. Оптимизируй это (Часть 1) Refactor Доработки Отправка на проверку
