+++ project/src/components/locations-list/locations-list.tsx
@@ -4,9 +4,8 @@ import {nanoid} from 'nanoid';
 import React from 'react';
 import {Dispatch} from 'redux';
 import {connect, ConnectedProps, useDispatch, useSelector} from 'react-redux';
-import {ActionsType} from '../../types/action';
 import {City} from '../../types/city';
-import {RootState, selectCity} from '../../store/new-reducer';
+import {RootState, selectCity} from '../../store/reducer';
 import {useAppSelector} from '../../hooks/useAppSelector';
 import { useAppDispatch } from '../../hooks/useAppDispatch';
