+++ project/src/components/submit-form-comment/submit-form-comment.tsx
@@ -1,8 +1,8 @@
 import {ReactComponent as IconStar} from '../../static/icon-star.svg';
 import {nanoid} from 'nanoid';
 import React, {FormEvent, useEffect, useState} from 'react';
-import {useSubmitCommentMutation} from '../../services/apiAxios';
-import {setCurrentOfferComments} from '../../store/new-reducer';
+import {useSubmitCommentMutation} from '../../services/api';
+import {setCurrentOfferComments} from '../../store/reducer';
 import {useAppDispatch} from '../../hooks/useAppDispatch';
 
 type OutsideCommentFormProps = {
