+++ project/src/hooks/useAppSelector.ts
@@ -1,4 +1,4 @@
 import {TypedUseSelectorHook, useSelector} from 'react-redux';
-import {RootState} from '../store/new-reducer';
+import {RootState} from '../store/reducer';
 
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
