+++ project/src/index.tsx
@@ -3,19 +3,19 @@ import ReactDOM from 'react-dom';
 import App from './components/app';
 import {Provider} from 'react-redux';
 import {configureStore} from '@reduxjs/toolkit';
-import appReducer from './store/new-reducer';
-import {apiAxios} from './services/apiAxios';
+import appReducer from './store/reducer';
+import {api} from './services/api';
 import {redirect} from './store/middlewares/redirect';
 
 export const store = configureStore({
   reducer: {
     app: appReducer,
-    [apiAxios.reducerPath]: apiAxios.reducer,
+    [api.reducerPath]: api.reducer,
   },
   middleware: (gDM) => gDM({
     thunk: {
-      extraArgument: apiAxios,
-    },}).concat(redirect, apiAxios.middleware),
+      extraArgument: api,
+    },}).concat(redirect, api.middleware),
 })
 
 ReactDOM.render(
