--- project/src/services/apin.ts
@@ -1,49 +1,0 @@
-import axios, {AxiosInstance, AxiosRequestConfig, AxiosResponse, AxiosError} from 'axios';
-import {getToken} from './token';
-
-const BASE_URL = 'https://8.react.pages.academy/six-cities';
-const REQUEST_TIMEOUT = 5000;
-
-enum HttpCode {
-  Unauthorized = 401,
-}
-
-type UnauthorizedCallback = () => void;
-type OnChangeLoadingStatus = () => void;
-
-const createAPIN = (): AxiosInstance => {
-  const api = axios.create({
-    baseURL: BASE_URL,
-    timeout: REQUEST_TIMEOUT,
-  });
-
-  api.interceptors.response.use(
-    (response: AxiosResponse) => {
-      // onChangeLoadingStatus();
-      return response;
-    },
-
-    (error: AxiosError) => {
-      const {response} = error;
-      // onChangeLoadingStatus();
-      if (response?.status === HttpCode.Unauthorized) {
-        // onUnauthorized();
-      }
-      return Promise.reject(error);
-    },
-  );
-
-  api.interceptors.request.use(
-    (config: AxiosRequestConfig) => {
-      const token = getToken();
-
-      if (token && config.headers) {
-        config.headers['x-token'] = token;
-      }
-      return config;
-    },
-  );
-  return api;
-};
-
-export {createAPIN};
