--- project/src/types/fetchOfferData.ts
@@ -1,3 +1,0 @@
-export type FetchOfferData = {
-  id: string;
-}
